# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FlyRank Auth API",
    "version": "1.0.0",
    "description": "Secure authentication API for FlyRank Backend Track Week 2 Assignment A4. Supabase Auth manages accounts and JWTs; the Express API verifies bearer tokens before protected routes run."
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "tags": [
    { "name": "Auth", "description": "Authentication and session endpoints" },
    { "name": "Public", "description": "Open endpoints" },
    { "name": "Protected", "description": "Endpoints requiring a valid bearer token" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the Supabase access token only. Swagger sends it as Authorization: Bearer <token>."
      }
    },
    "schemas": {
      "Credentials": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "test@example.com" },
          "password": { "type": "string", "format": "password", "example": "password123" }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": ["refresh_token"],
        "properties": { "refresh_token": { "type": "string" } }
      },
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Create a user account",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Credentials" } } } },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Missing or invalid input" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Authenticate and return access and refresh tokens",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Credentials" } } } },
        "responses": {
          "200": { "description": "Login successful" },
          "400": { "description": "Missing input" },
          "401": { "description": "Invalid login credentials" },
          "429": { "description": "Too many failed login attempts" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Exchange a refresh token for a fresh access token",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } } } },
        "responses": {
          "200": { "description": "New access token issued" },
          "400": { "description": "Missing refresh token" },
          "401": { "description": "Invalid or expired refresh token" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "End the current authenticated session",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logged out" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "tags": ["Protected"],
        "summary": "Read safe current-user profile metadata",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Authenticated user metadata" },
          "401": { "description": "Missing, malformed, invalid, or expired token" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "tags": ["Protected"],
        "summary": "Second protected route using the same middleware",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Dashboard data" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/protected/admin": {
      "get": {
        "tags": ["Protected"],
        "summary": "Stretch: admin-only route demonstrating 403",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Admin access" }, "401": { "description": "Unauthorized" }, "403": { "description": "Authenticated but not authorized" } }
      }
    },
    "/public/info": {
      "get": {
        "tags": ["Public"],
        "summary": "Read public information",
        "responses": { "200": { "description": "Public information" } }
      }
    }
  }
}